initial public working commit (partially styled)
The first lines of code on this project were written in March 2015, which was CSON data files.

In January 2017 it was revised with data verification.

In February 2017 a react interface was added.

The project was halted as Auth0 proved too difficult to adopt.

In April 2019 the project was kicked off again using lit-html instead of react, and cloudflare access instead of auth0, it was a success, hence this commit.